requirements.txt
pyserial
pandas
matplotlib
seaborn
scipy
